refactoring mappers
