+++ lib/data/repository/task_repository.dart
@@ -1,6 +1,5 @@
 import 'package:clean_flutter_app/data/cache/source/task_cds.dart';
-import 'package:clean_flutter_app/data/mapper/cache_to_domain.dart';
-import 'package:clean_flutter_app/data/mapper/domain_to_cache.dart';
+import 'package:clean_flutter_app/data/mapper/task_mappers.dart';
 import 'package:domain/data_repository/task_repository.dart';
 import 'package:domain/model/task.dart';
 import 'package:domain/model/task_list_orientation.dart';
